+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.16.0
 * Can use ListAdapter as "content" property value of ListLayout
 * The IsListItemEnabled method of the ListAdapter interface has been made optional 
+* Can use GridAdapter as "content" property value of GridLayout
 * Bug fixing
 
 # v0.15.0